Conflicts:
	src/tests/SystemTest.java

+++ src/Store/IProduct.java
@@ -6,7 +6,7 @@ public interface IProduct {
     public String getName();
     public String getCategory();
     public List<String> getKeyWords();
-    public int getPrice();
+    public double getPrice();
     public int getRating();
     public Store getStore();
 }
